Zach/gameview

+++ app/src/main/java/com/d/tickettoride/model/gameplay/City.kt
@@ -1,3 +1,3 @@
 package com.d.tickettoride.model.gameplay
 
-data class City(val id:Int, val latitude:Double, val longitude:Double, val name:String)
+data class City(val id:Int, val latitude:Float, val longitude:Float, val name:String)
